State for supervisor page
